test: 게임 상태 테스트 분리 및 폴더 생성
